Fix broken build, add test, rename to scalapb_proto_library
