+++ test_run.sh
@@ -582,3 +582,4 @@ $runner test_scala_library_expect_no_recompilation_on_internal_change_of_scala_d
 $runner test_scala_library_expect_no_recompilation_on_internal_change_of_java_dependency
 $runner test_scala_library_expect_no_java_recompilation_on_internal_change_of_scala_sibling
 $runner test_scala_library_expect_failure_on_missing_direct_java
+$runner bazel run test:test_scala_proto_server
